Move testing plugins into internal/testutil
Moves the stub plugins into a dedicated package for each plugin inside
`internal/testutil/plugin`.
